docker/institution_service/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
djangorestframework>=3.14
gunicorn>=21.2
requests
